Helper API changed, now the plugin is chosen in the main application and transmitted along with the ROM path when the main app setup the emulation. The emulation itself is started after the GameCore has been added as observer for its settings.
git-svn-id: https://openemu.svn.sourceforge.net/svnroot/openemu/branches/iosurface@711 a16923d9-94c8-4a39-ae78-11df2d60a2ba

+++ OpenEmu/TaskWrapper.h
@@ -58,39 +58,40 @@
 
 // Your controller's implementation of this method will be called when output arrives from the NSTask.
 // Output will come from both stdout and stderr, per the TaskWrapper implementation.
-- (void)appendOutput:(NSString *)output fromProcess: (TaskWrapper *)aTask;
+- (void)appendOutput:(NSString *)output fromProcess:(TaskWrapper *)aTask;
 
 // This method is a callback which your controller can use to do other initialization when a process
 // is launched.
-- (void)processStarted: (TaskWrapper *)aTask;
+- (void)processStarted:(TaskWrapper *)aTask;
 
 // This method is a callback which your controller can use to do other cleanup when a process
 // is halted.
-- (void)processFinished: (TaskWrapper *)aTask withStatus: (int)statusCode;
+- (void)processFinished:(TaskWrapper *)aTask withStatus:(NSInteger)statusCode;
 
 @end
 
-@interface TaskWrapper : NSObject {
-    NSTask 			*task;
-    id				<TaskWrapperController>controller;
-	id				userInfo;
-    NSArray			*arguments;
+@interface TaskWrapper : NSObject
+{
+    NSTask                    *task;
+    id<TaskWrapperController>  controller;
+    id                         userInfo;
+    NSArray                   *arguments;
 }
 
-@property (readonly) NSTask* task;
+@property(readonly) NSTask *task;
 
 // This is the designated initializer - pass in your controller and any task arguments.
 // The first argument should be the path to the executable to launch with the NSTask.
-- (id)initWithController:(id <TaskWrapperController>)controller arguments:(NSArray *)args userInfo: (id)someInfo;
+- (id)initWithController:(id<TaskWrapperController>)controller arguments:(NSArray *)args userInfo:(id)someInfo;
 
 // This method launches the process, setting up asynchronous feedback notifications.
-- (void) startProcess;
+- (void)startProcess;
 
 // This method stops the process, stoping asynchronous feedback notifications.
-- (void) stopProcess;
+- (void)stopProcess;
 
 // This is to send some data to the process
-- (void)sendToProcess: (NSString *)aString;
+- (void)sendToProcess:(NSString *)aString;
 
 // External check if we're still running
 - (BOOL)isRunning;
